// src/api/generated/models/leagueCategoryPriceResource.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export interface LeagueCategoryPriceResource {
  id: string;
  league_id: string;
  category: string;
  min_overall: number;
  base_salary: string;
  base_passe: string;
  created_at: string;
  updated_at: string;
}